requirements.txt
pandas
numpy
pyarrow
mlflow
azure-storage-blob
scikit-learn
# optional: for optimization wrapper
pulp
